Fixing some compiler warnings by adding some casts, and commenting out unused parameters / renaming conflicting parameters.

+++ examples/ARA/ARASampleProject/Source/RegionSequenceView.cpp
@@ -44,7 +44,7 @@ void RegionSequenceView::paint (Graphics& g)
     g.addTransform (AffineTransform::rotation (-MathConstants<float>::halfPi, 
                                                headerRect.getWidth () * 0.5f,
                                                headerRect.getHeight () * 0.5f));
-    Rectangle<int> textRect (-headerRect.getHeight (), 1.25 * headerRect.getWidth (), 1.5 * headerRect.getHeight (), headerRect.getWidth ());
+    Rectangle<int> textRect (-headerRect.getHeight (), (int)(1.25 * headerRect.getWidth ()), (int)(1.5 * headerRect.getHeight ()), headerRect.getWidth ());
     g.setColour (trackColour.contrasting (1.0f));
     g.setFont (Font (12.0));
     g.drawText ("Track #" + String (regionSequence->getOrderIndex()) + ": " + regionSequence->getName(), textRect, Justification::bottomLeft);
